fix: 继续优化延迟问题

+++ xiaomusic/cli.py
@@ -143,39 +143,48 @@ def main():
     except Exception as e:
         print(f"Execption {e}")
 
-    def run_server():
+    def run_server(port):
         xiaomusic = XiaoMusic(config)
         HttpInit(xiaomusic)
         uvicorn.run(
             HttpApp,
-            host="127.0.0.1",
-            port=config.port + 1,
+            host=["0.0.0.0", "::"],
+            port=port,
             log_config=LOGGING_CONFIG,
         )
 
-    command = [
-        "uvicorn",
-        "xiaomusic.gate:app",
-        "--workers",
-        "4",
-        "--host",
-        "0.0.0.0",
-        "--port",
-        str(config.port),
-    ]
+    process = None
 
-    process = subprocess.Popen(command)
+    def run_gate():
+        command = [
+            "uvicorn",
+            "xiaomusic.gate:app",
+            "--workers",
+            "4",
+            "--host",
+            "0.0.0.0",
+            "--port",
+            str(config.port),
+        ]
+        global process
+        process = subprocess.Popen(command)
+
     def signal_handler(sig, frame):
         print("主进程收到退出信号，准备退出...")
-        process.terminate()  # 终止子进程
-        process.wait()  # 等待子进程退出
-        print("子进程已退出")
+        if process is not None:
+            process.terminate()  # 终止子进程
+            process.wait()  # 等待子进程退出
+            print("子进程已退出")
         os._exit(0)  # 退出主进程
 
     # 捕获主进程的退出信号
     signal.signal(signal.SIGINT, signal_handler)
     signal.signal(signal.SIGTERM, signal_handler)
-    run_server()
+    if config.enable_gate:
+        run_gate()
+        run_server(config.port + 1)
+    else:
+        run_server(config.port)
 
 
 if __name__ == "__main__":
